visualizer.lua: improve compatiblity with other scripts
Writing to lavfi-complex even when no visualization is desired can cause
conflicts with other scripts, as can be seen in
https://github.com/mfcc64/mpv-scripts/issues/36

Since each file needs to be evaluated individually if the visualizer
should be used for it. There is no benefit in keeping lavfi-complex set
across file boundaries and only adds the requirement that it has to be
reset if the next file shouldn't use the visualizer.
Setting lavfi-complex as a file-local-option avoids the need to reset it
via the script, and can help avoid conflicts with other scripts.

Additionally using the visualizer for files without audio doesn't make
sense, therefore the visualizer can be skipped for any files without
audio tracks.