feat: update visualizer when opts changed on runtime
The `options.read_options` will not, by default, update changes if opts
changed on runtime. Quoting from the man pages:

> The `on_update` parameter  enables run-time updates of all matching
option values via the script-opts option/property. If any of the
matching options changes, the values in the table (which was originally
passed to the function) are changed, and on_update(list) is called.

With this commit, the visualizer will now update when the opts changed
on runtime. This will also allow user to, for example, create a keybind
that will toggle the visualizer on and off, *not cycling them*. Like so:
```
c cycle-values script-opts "visualizer-name=off" "visualizer-name=showcqtbar"
```